requirements.txt
fastapi==0.128.4
sqlalchemy==2.0.46
dotenv==0.9.9
pydantic_settings==2.12.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
asyncpg==0.31.0
uvicorn==0.40.0
grpcio==1.78.0
grpcio-tools==1.78.0
redis==7.2.1
python-jose[cryptography]==3.5.0